+++ R/copy_if_not_already_exist.R
@@ -26,10 +26,7 @@ copy_if_not_already_exist <- function(
 ){
 
   # list extensions in input dir
-  ext_list <- setdiff(
-    dir_ls(path = from, recurse = 1, type = "directory"),
-    dir_ls(path = from, recurse = 0, type = "directory")
-  ) |> 
+  ext_list <- dir_ls(path = from, recurse = 1, type = "directory") |> 
     gsub(pattern = from, replacement = "")
   
   # list extensions already present in target dir
@@ -42,7 +39,7 @@ copy_if_not_already_exist <- function(
       cli_alert_info(
         paste(
           "{toString(existing_ext)} extension(s)",
-          "already present in quarto project {dirname(from)}.",
+          "already present in quarto project {dirname(to)}.",
           "Using it for compil."
         )
       )