+++ R/render_single_qmd.R
@@ -44,12 +44,8 @@
 render_single_qmd <- function(
     qmd,
     img_root_dir = "img",
-    output_format = "thinkridentity-revealjs",
-    metadata = list(
-      footer = "",
-      logo = "",
-      "title-slide-attributes" = list("data-background-image" = "")
-    )
+    output_format = "revealjs",
+    metadata = NULL
 ) {
   # set image sub-folder name
   chapter <- dirname(qmd)